refactor: ReComment entity -> Comment entity로 병합
- ReComment 관련 클래스, 메서드, 변수 삭제
- 필요없는 impot 삭제

- fix: CommentController Ambigious error 대댓글 생성 컨트롤러 매핑 수정
