--- src/main/java/com/donguri/jejudorang/domain/community/repository/comment/ReCommentRepository.java
@@ -1,7 +1,0 @@
-package com.donguri.jejudorang.domain.community.repository.comment;
-
-import com.donguri.jejudorang.domain.community.entity.comment.ReComment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReCommentRepository extends JpaRepository<ReComment, Long> {
-}
