+++ src/main/java/com/donguri/jejudorang/domain/community/service/comment/CommentService.java
@@ -2,8 +2,8 @@ package com.donguri.jejudorang.domain.community.service.comment;
 
 import com.donguri.jejudorang.domain.community.dto.request.comment.CommentRequest;
 import com.donguri.jejudorang.domain.community.dto.request.comment.CommentRequestWithId;
+import com.donguri.jejudorang.domain.community.dto.request.comment.ReCommentRequest;
 import com.donguri.jejudorang.domain.community.dto.response.comment.CommentResponse;
-import org.springframework.data.domain.Page;
 
 import java.util.List;
 
@@ -11,6 +11,9 @@ public interface CommentService {
 
     // 새 댓글 작성
     void writeNewComment(String accessToken, Long postId, CommentRequest newComment);
+
+    // 새 대댓글 작성
+    void writeNewReComment(String accessToken, ReCommentRequest newReComment);
 
     // 댓글 수정
     void modifyComment(String accessToken, CommentRequestWithId commentToUpdate) throws IllegalAccessException;
